requirements.txt
snowflake-connector-python
snowflake-snowpark-python[pandas]
snowflake-sqlalchemy
langchain
langgraph
langchain_community
langchain[openai]
langchain[anthropic]
langchain[google_genai]
openai
langchain-experimental
pydantic
pandas
streamlit
python-telegram-bot
python-dotenv
flask
pywa 
tzdata
timezonefinder
streamlit-js-eval
